# REST/good-notes/waterloo/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Doctors Scheduler",
      "version": "1.0.0"
    },
    "schemes": [
      "http",
      "https"
    ],
    "basePath": "/api/v1",
    "paths": {
...
      "/doctors/{id}": {
        "get": {
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "type": "integer",
              "required": true
            }
          ],
          "summary": "Get single doctor",
          "description": "Get doctor by id",
          "responses": {
            "200": {
              "description": "Single doctor",
              "schema": {
                "$ref": "#/definitions/Doctor"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
      "Doctor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }